changed everything to double

+++ MidtermProject/denseGEMM-SUMMA/include/utils.h
@@ -2,9 +2,9 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-void matmul(float* A, float* B, float* C, int m, int n, int k);
-void verify_result(float* C_global, float* A, float* B, int m, int n, int k);
-float* generate_matrix_A(int rows, int cols, int rank);
-float* generate_matrix_B(int rows, int cols, int rank);
+void matmul(double *A, double *B, double *C, int m, int n, int k);
+void verify_result(double *C_global, double *A, double *B, int m, int n, int k);
+double *generate_matrix_A(int rows, int cols, int rank);
+double *generate_matrix_B(int rows, int cols, int rank);
 
 #endif
